requirements.txt
marker-pdf
torch
transformers
pdf2image
uvicorn
fastapi
python-multipart
openai
google-generativeai
boto3
groq
pydantic-settings
PyYAML
python-dotenv